Circus: Get rid of more Windows typedefs

+++ Extract/Circus.h
@@ -11,15 +11,15 @@ public:
 private:
 	struct SPCMHeader
 	{
-		char  identifier[4]; // Identifier
-		DWORD data_size;     // Data Size
-		DWORD flags;         // Flags
-		WORD  format_id;     // Format ID
-		WORD  channels;      // Number of channels
-		DWORD freq;          // Sampling rate
-		DWORD bps;           // Data Rate
-		WORD  block_size;    // Block size
-		WORD  bits;          // Number of bits
+		char identifier[4]; // Identifier
+		u32  data_size;     // Data Size
+		u32  flags;         // Flags
+		u16  format_id;     // Format ID
+		u16  channels;      // Number of channels
+		u32  freq;          // Sampling rate
+		u32  bps;           // Data Rate
+		u16  block_size;    // Block size
+		u16  bits;          // Number of bits
 	};
 
 	bool MountPCK(CArcFile* archive);
@@ -28,8 +28,8 @@ private:
 	bool MountPCM(CArcFile* archive);
 
 	bool DecodeCRX(CArcFile* archive);
-	bool DecodeCRX1(CArcFile* archive, const u8* src, DWORD src_size);
-	bool DecodeCRX2(CArcFile* archive, const u8* src, DWORD src_size);
+	bool DecodeCRX1(CArcFile* archive, const u8* src, size_t src_size);
+	bool DecodeCRX2(CArcFile* archive, const u8* src, size_t src_size);
 	bool DecodePCM(CArcFile* archive);
 	bool DecodePCM1(CArcFile* archive, const SPCMHeader& pcm_header);
 	bool DecodePCM2(CArcFile* archive, const SPCMHeader& pcm_header);
